rewrite ghidra sync. using mapper inside manager

+++ SDA/Module/DB/Mappers/ClassTypeMapper.cpp
@@ -94,6 +94,8 @@ void ClassTypeMapper::doUpdate(TransactionContext* ctx, IDomainObject* obj)
 
 void ClassTypeMapper::doRemove(TransactionContext* ctx, IDomainObject* obj)
 {
+	if (ctx->m_notDelete)
+		return;
 	SQLite::Statement query(*ctx->m_db, "DELETE FROM sda_classes WHERE struct_id=?1");
 	query.bind(1, obj->getId());
 	query.exec();
